Refactoring + Passing test. TODO: Error simulation
