+++ Makefile
@@ -91,7 +91,7 @@ fmt: ## Run go fmt against code.
 
 .PHONY: vet
 vet: ## Run go vet against code.
-	go vet --tags e2e,istio,telemetrycomponents ./...
+	go vet --tags e2e,istio ./...
 
 .PHONY: tidy
 tidy: ## Check if there any dirty change for go mod tidy.
